Move CompleteProfileWithSocial out of desktop dir
